Refactor save as dialog / field.
