+++ public/static/locales/en/common.json
@@ -38,6 +38,7 @@
     "closeMenu": "Close Menu",
     "refGenomes": "Reference Genomes",
     "refresh": "Refresh",
+    "required": "Required.",
     "save": "Save",
     "saveAs": "Save As",
     "search": "Search...",
