+++ src/components/data/SaveAsField.js
@@ -3,7 +3,7 @@
  *
  * A field to collect path (with filename) from a user to save a newly created file.
  */
-import React from "react";
+import React, { useEffect, useState } from "react";
 
 import { useTranslation } from "i18n";
 
@@ -11,80 +11,80 @@ import ids from "./ids";
 import styles from "./styles";
 import withErrorAnnouncer from "components/utils/error/withErrorAnnouncer";
 
-import {
-    build as buildDebugId,
-    FormTextField,
-    getFormError,
-} from "@cyverse-de/ui-lib";
+import DataSelectionDrawer from "components/data/SelectionDrawer";
+import { UploadTrackingProvider } from "contexts/uploadTracking";
+import { useUserProfile } from "contexts/userProfile";
+import { useConfig } from "contexts/config";
+import ResourceTypes from "components/models/ResourceTypes";
+import { parseNameFromPath } from "./utils";
+
+import { build as buildDebugId, FormTextField } from "@cyverse-de/ui-lib";
 
 import {
     Button,
+    CircularProgress,
     Grid,
+    Hidden,
+    Paper,
     IconButton,
-    InputAdornment,
     makeStyles,
+    Typography,
+    useTheme,
 } from "@material-ui/core";
 
 import ClearIcon from "@material-ui/icons/Clear";
-import SaveAsDialog from "./SaveAsDialog";
+import { Folder } from "@material-ui/icons";
 
 const useStyles = makeStyles(styles);
-
-const SaveAsButton = (props) => {
-    const { baseId, startingPath, onConfirm } = props;
-    const { t } = useTranslation("data");
-    const [open, setOpen] = React.useState(false);
-
-    return (
-        <>
-            <Button
-                id={buildDebugId(baseId, ids.SAVE_AS_BTN)}
-                color="primary"
-                size="small"
-                variant="outlined"
-                onClick={() => setOpen(true)}
-            >
-                {t("saveas")}
-            </Button>
-            <SaveAsDialog
-                path={startingPath}
-                open={open}
-                onClose={() => setOpen(false)}
-                onSaveAs={(path) => {
-                    onConfirm(path);
-                    setOpen(false);
-                }}
-            />
-        </>
-    );
-};
 
 const SaveAsField = ({ startingPath, showErrorAnnouncer, ...props }) => {
     // These props need to be spread down into the FormTextField
     const {
         id,
+        path,
         field: { name, value },
-        form: { touched, errors, setFieldValue },
+        form: { setFieldValue },
         required,
+        loading,
     } = props;
+
     const classes = useStyles();
+    const theme = useTheme();
     const { t } = useTranslation("data");
-    const errorMsg = getFormError(name, touched, errors);
+    const [userProfile] = useUserProfile();
+    const [config] = useConfig();
+    const username = userProfile?.id;
+    const irodsHomePath = config?.irods?.home_path;
+    const [dest, setDest] = useState(path || `${irodsHomePath}/${username}`);
+    const [selectionDrawerOpen, setSelectionDrawerOpen] = useState(false);
+
+    const baseId = "saveAsFld";
+
+    useEffect(() => {
+        const val = parseNameFromPath(value);
+        if (val) {
+            setFieldValue(name, `${dest}/${val}`);
+        } else {
+            setFieldValue(name, "");
+        }
+    }, [setFieldValue, dest, value, name]);
 
     const inputProps = {
-        readOnly: true,
+        readOnly: loading,
         endAdornment: (
-            <InputAdornment position="end">
-                <SaveAsButton
-                    baseId={id}
-                    startingPath={startingPath}
-                    multiSelect={false}
-                    name={name}
-                    onConfirm={(path) => {
-                        setFieldValue(name, path);
-                    }}
-                />
-            </InputAdornment>
+            <>
+                {loading && (
+                    <CircularProgress
+                        id={buildDebugId(
+                            baseId,
+                            ids.FOLDER_NAME,
+                            ids.LOADING_SKELETON
+                        )}
+                        color="inherit"
+                        size={20}
+                    />
+                )}
+            </>
         ),
     };
 
@@ -102,25 +102,67 @@ const SaveAsField = ({ startingPath, showErrorAnnouncer, ...props }) => {
     }
 
     return (
-        <Grid
-            container
-            direction="row"
-            justify="flex-start"
-            alignItems="center"
-        >
-            <Grid item xs>
-                <FormTextField
-                    InputProps={inputProps}
-                    size="small"
-                    className={classes.inputSelectorTextFiled}
-                    error={errorMsg !== null}
-                    {...props}
-                />
+        <Paper style={{ padding: theme.spacing(0.5) }}>
+            <Grid
+                container
+                direction="column"
+                justify="center"
+                alignItems="stretch"
+                spacing={1}
+            >
+                <Grid item xs>
+                    <Typography
+                        classes={{
+                            root: classes.bottomPadding,
+                        }}
+                        component="span"
+                        variant="body2"
+                    >
+                        {t("newFileLocation", {
+                            path: dest,
+                        })}
+                    </Typography>
+                    <Button
+                        id={buildDebugId(baseId, ids.SAVE_BTN)}
+                        size="small"
+                        style={{
+                            marginLeft: theme.spacing(1),
+                        }}
+                        variant="outlined"
+                        disableElevation
+                        color="primary"
+                        onClick={() => setSelectionDrawerOpen(true)}
+                        startIcon={<Folder fontSize="small" />}
+                    >
+                        <Hidden xsDown>{t("browse")}</Hidden>
+                    </Button>
+                </Grid>
+                <Grid item xs>
+                    <FormTextField
+                        id={buildDebugId(baseId, id)}
+                        size="small"
+                        className={classes.inputSelectorTextFiled}
+                        InputProps={inputProps}
+                        {...props}
+                    />
+                </Grid>
             </Grid>
-        </Grid>
+            <UploadTrackingProvider>
+                <DataSelectionDrawer
+                    open={selectionDrawerOpen}
+                    startingPath={dest}
+                    acceptedType={ResourceTypes.Folder}
+                    onClose={() => setSelectionDrawerOpen(false)}
+                    onConfirm={(selection) => {
+                        setDest(selection);
+                        setSelectionDrawerOpen(false);
+                    }}
+                    baseId={buildDebugId(baseId, "dataSelection")}
+                    multiSelect={false}
+                />
+            </UploadTrackingProvider>
+        </Paper>
     );
 };
-
-export { SaveAsButton };
 
 export default withErrorAnnouncer(SaveAsField);
